refactor: gameover 페이지로 분리
